rag-demo-be/requirements.txt
flask
flask-cors
python-dotenv

langchain
langchain-community
langchain-huggingface
langchain-text-splitters

chromadb
sentence-transformers
pypdf

huggingface-hub
redis
